Completed credits page UI
